move form submit handlers to their own file
handle clearing the selected auto link
